rename table origin to Last Component Update info
