+++ src/scripts/react/common/TableUpdatedByComponentInfo.jsx
